17-wildlife-classifier/pyproject.toml
[project]
name = "wild-eye"
version = "0.1.0"
description = "WildEye - Wildlife Species Classifier from camera trap images"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "WildEye Team" }]
keywords = ["wildlife", "conservation", "camera-trap", "species-classifier", "biodiversity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "boto3>=1.34.0",
    "pydantic>=2.6.0",
    "streamlit>=1.31.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    "python-multipart>=0.0.9",
    "exifread>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "boto3-stubs[dynamodb,s3]>=1.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wild_eye"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring AWS resources",
]
